Em's change
So, I'm a little confused about what this code pattern is showing. I originally thought you were going to build an app, but it seems like what you're doing is connecting a sample movie ticket reservation app to a Crunchy DB. Is that correct? 

I left a few other questions in there with <!--EM. Please respond to those and we'll go from there.